requirements.txt
aiohttp
aiofiles
asyncio
APScheduler==3.8.1
ffmpeg-python
pillow
pyrogram==1.4.16
py-tgcalls==0.8.6
pytz==2021.3
pytube
wget
python-dotenv
search-engine-parser
tgcrypto
youtube-search-python
youtube-search
speedtest-cli
cryptography
dnspython
future
googletrans
hurry.filesize
motor
psutil
pykeyboard
pyromod
requests
speedtest-cli
uvloop
yt-dlp==2021.11.10.1
telegraph
hachoir
lyricsgenius
httpx>=0.14.2
